pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "simace"
dynamic = ["version"]
description = "ACE pedigree simulation with A/C/E variance components"
requires-python = ">=3.13"
dependencies = [
    "pedigree-graph @ git+https://github.com/rwaples/pedigree-graph.git@v0.2.0",
    "numpy>=2.2,<3",
    "scipy>=1.14,<2",
    "pandas>=3.0,<4",
    "pyarrow>=17,<24",
    "pyyaml>=6,<7",
    "numba>=0.60,<1",
]

[project.optional-dependencies]
plot = [
    "matplotlib>=3.9,<4",
    "seaborn>=0.13,<1",
    "pillow>=10,<14",
]
test = [
    "pytest>=9,<10",
    "pytest-cov>=7,<8",
    "ruff>=0.8",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.27",
]
workflow = [
    "snakemake>=9,<10",
    "snakefmt>=1,<2",
]
dev = [
    "simace[plot,test,docs,workflow]",
]

[project.scripts]
simace-simulate = "simace.simulation.simulate:cli"
simace-phenotype = "simace.phenotyping.phenotype:cli"
simace-threshold = "simace.phenotyping.threshold:cli"
simace-sample = "simace.sampling.sample:cli"
simace-validate = "simace.analysis.validate:cli"
simace-phenotype-stats = "simace.analysis.stats:cli"
simace-gather = "simace.analysis.gather:cli"
simace-plot-phenotype = "simace.plotting.plot_phenotype:cli"
simace-plot-validation = "simace.plotting.plot_validation:cli"
simace-parquet-to-tsv = "simace.core.parquet_to_tsv:cli"

[tool.setuptools.packages.find]
include = ["simace*"]

[tool.setuptools_scm]

[tool.pytest.ini_options]
testpaths = ["tests"]
norecursedirs = ["fitACE", ".git", ".snakemake", "external", "site"]
markers = [
    "slow: heavyweight tests (Monte Carlo, large pedigrees); deselect with -m 'not slow'",
]

[tool.ruff]
target-version = "py313"
line-length = 120
extend-exclude = [".snakemake", "notes", "fitACE"]

[tool.ruff.lint]
select = ["F", "E", "W", "I", "UP", "B", "SIM", "RUF", "A", "C4", "PT", "PIE", "PERF", "TCH", "RSE", "RET", "D"]
ignore = [
    "E501",   # line too long — formatter handles wrapping
    "SIM108", # use ternary — often less readable
    "RUF002", # ambiguous unicode in docstring — scientific notation (×)
    "RUF003", # ambiguous unicode in comment — scientific notation (×)
    "C408",   # unnecessary dict() call — dict() more readable in matplotlib bbox= kwargs
    "D105",   # undocumented magic method — not needed
    "D107",   # undocumented __init__ — args documented in class docstring
]

[tool.ruff.lint.per-file-ignores]
"workflow/scripts/**/*.py" = ["F821", "B018"]  # snakemake global injected at runtime; bare name check idiom
"tests/*.py" = ["B011", "RUF012", "D"]       # assert False; mutable class attrs OK in tests; no docstrings needed
"tests/**/*.py" = ["D"]                      # no docstrings needed in tests
"simace/__init__.py" = ["F401"]             # setup_logging used via re-export
"simace/core/_numba_utils.py" = ["D"]               # private module, all functions private
"simace/phenotyping/models/*.py" = ["D102"]         # method docstrings inherited from PhenotypeModel ABC
"simace/plotting/atlas_manifest.py" = ["RUF001"]    # caption strings use scientific × − – etc.

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["simace"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
